requirements.txt
torch
torchvision
torchaudio
torch_geometric
pandas
numpy
PyYAML
omegaconf
tqdm
matplotlib
umap-learn
geomloss
scipy
torchdiffeq